+++ test/main/Test/Format/Format.hs
@@ -10,12 +10,12 @@ import Test.Tasty.HUnit
 import Test.TestUtil
 
 -- as found in http://www.opengroup.org/onlinepubs/007908799/xsh/strftime.html
--- plus FgGklz
+-- plus FgGklvz
 -- f not supported
 -- P not always supported
 -- s time-zone dependent
 chars :: [Char]
-chars = "aAbBcCdDeFgGhHIjklmMnprRStTuUVwWxXyYzZ%"
+chars = "aAbBcCdDeFgGhHIjklmMnprRStTuUvVwWxXyYzZ%"
 
 -- as found in "man strftime" on a glibc system. '#' is different, though
 modifiers :: [Char]
@@ -115,103 +115,23 @@ testNominalDiffTime =
     "NominalDiffTime"
     []
 
-{-
-        [ testAFormat "%ww%Dd%Hh%Mm%ESs" "3w2d2h22m8.21s" $ (fromRational $ 23 * 86400 + 8528.21 :: NominalDiffTime)
-        , testAFormat "%dd %hh %mm %ss %Ess" "0d 0h 0m 0s 0.74s" $ (fromRational $ 0.74 :: NominalDiffTime)
-        , testAFormat "%dd %hh %mm %ss %Ess" "0d 0h 0m 0s -0.74s" $ (fromRational $ negate $ 0.74 :: NominalDiffTime)
-        , testAFormat "%dd %hh %mm %ss %Ess %0Ess" "23d 554h 33262m 1995728s 1995728.21s 1995728.210000000000s" $
-            (fromRational $ 23 * 86400 + 8528.21 :: NominalDiffTime)
-        , testAFormat "%ww%Dd%Hh%Mm%Ss" "-3w-2d-2h-22m-8s" $
-            (fromRational $ negate $ 23 * 86400 + 8528.21 :: NominalDiffTime)
-        , testAFormat "%ww%Dd%Hh%Mm%ESs" "-3w-2d-2h-22m-8.21s" $
-            (fromRational $ negate $ 23 * 86400 + 8528.21 :: NominalDiffTime)
-        , testAFormat "%ww%Dd%Hh%Mm%Ss" "-3w-2d-2h-22m0s" $
-            (fromRational $ negate $ 23 * 86400 + 8520.21 :: NominalDiffTime)
-        , testAFormat "%ww%Dd%Hh%Mm%ESs" "-3w-2d-2h-22m-0.21s" $
-            (fromRational $ negate $ 23 * 86400 + 8520.21 :: NominalDiffTime)
-        , testAFormat "%dd %hh %mm %Ess" "-23d -554h -33262m -1995728.21s" $
-            (fromRational $ negate $ 23 * 86400 + 8528.21 :: NominalDiffTime)
-        , testAFormat "%3Es" "1.200" (1.2 :: NominalDiffTime)
-        , testAFormat "%3ES" "01.200" (1.2 :: NominalDiffTime)
-        , testAFormat "%3ES" "01.200" (61.2 :: NominalDiffTime)
-        , testAFormat "%3Es" "1.245" (1.24582 :: NominalDiffTime)
-        , testAFormat "%3ES" "01.245" (1.24582 :: NominalDiffTime)
-        , testAFormat "%3ES" "01.245" (61.24582 :: NominalDiffTime)
-        ]
--}
-
 testDiffTime :: TestTree
 testDiffTime =
   testGroup
     "DiffTime"
     []
 
-{-
-        [ testAFormat "%ww%Dd%Hh%Mm%ESs" "3w2d2h22m8.21s" $ (fromRational $ 23 * 86400 + 8528.21 :: DiffTime)
-        , testAFormat "%dd %hh %mm %ss %Ess" "0d 0h 0m 0s 0.74s" $ (fromRational $ 0.74 :: DiffTime)
-        , testAFormat "%dd %hh %mm %ss %Ess" "0d 0h 0m 0s -0.74s" $ (fromRational $ negate $ 0.74 :: DiffTime)
-        , testAFormat "%dd %hh %mm %ss %Ess %0Ess" "23d 554h 33262m 1995728s 1995728.21s 1995728.210000000000s" $
-            (fromRational $ 23 * 86400 + 8528.21 :: DiffTime)
-        , testAFormat "%ww%Dd%Hh%Mm%Ss" "-3w-2d-2h-22m-8s" $ (fromRational $ negate $ 23 * 86400 + 8528.21 :: DiffTime)
-        , testAFormat "%ww%Dd%Hh%Mm%ESs" "-3w-2d-2h-22m-8.21s" $
-            (fromRational $ negate $ 23 * 86400 + 8528.21 :: DiffTime)
-        , testAFormat "%ww%Dd%Hh%Mm%Ss" "-3w-2d-2h-22m0s" $ (fromRational $ negate $ 23 * 86400 + 8520.21 :: DiffTime)
-        , testAFormat "%ww%Dd%Hh%Mm%ESs" "-3w-2d-2h-22m-0.21s" $
-            (fromRational $ negate $ 23 * 86400 + 8520.21 :: DiffTime)
-        , testAFormat "%dd %hh %mm %Ess" "-23d -554h -33262m -1995728.21s" $
-            (fromRational $ negate $ 23 * 86400 + 8528.21 :: DiffTime)
-        , testAFormat "%3Es" "1.200" (1.2 :: DiffTime)
-        , testAFormat "%3ES" "01.200" (1.2 :: DiffTime)
-        , testAFormat "%3ES" "01.200" (61.2 :: DiffTime)
-        , testAFormat "%3Es" "1.245" (1.24582 :: DiffTime)
-        , testAFormat "%3ES" "01.245" (1.24582 :: DiffTime)
-        , testAFormat "%3ES" "01.245" (61.24582 :: DiffTime)
-        ]
--}
-
-testCalenderDiffDays :: TestTree
-testCalenderDiffDays =
+testCalendarDiffDays :: TestTree
+testCalendarDiffDays =
   testGroup
-    "CalenderDiffDays"
+    "CalendarDiffDays"
     []
 
-{-
-        [ testAFormat "%yy%Bm%ww%Dd" "5y4m3w2d" $ CalendarDiffDays 64 23
-        , testAFormat "%bm %dd" "64m 23d" $ CalendarDiffDays 64 23
-        , testAFormat "%yy%Bm%ww%Dd" "-5y-4m-3w-2d" $ CalendarDiffDays (-64) (-23)
-        , testAFormat "%bm %dd" "-64m -23d" $ CalendarDiffDays (-64) (-23)
-        ]
--}
-
-testCalenderDiffTime :: TestTree
-testCalenderDiffTime =
+testCalendarDiffTime :: TestTree
+testCalendarDiffTime =
   testGroup
-    "CalenderDiffTime"
+    "CalendarDiffTime"
     []
-
-{-
-        [ testAFormat "%yy%Bm%ww%Dd%Hh%Mm%Ss" "5y4m3w2d2h22m8s" $ CalendarDiffTime 64 $ 23 * 86400 + 8528.21
-        , testAFormat "%yy%Bm%ww%Dd%Hh%Mm%ESs" "5y4m3w2d2h22m8.21s" $ CalendarDiffTime 64 $ 23 * 86400 + 8528.21
-        , testAFormat "%yy%Bm%ww%Dd%Hh%Mm%0ESs" "5y4m3w2d2h22m08.210000000000s" $
-            CalendarDiffTime 64 $
-                23 * 86400 + 8528.21
-        , testAFormat "%bm %dd %hh %mm %Ess" "64m 23d 554h 33262m 1995728.21s" $
-            CalendarDiffTime 64 $
-                23 * 86400 + 8528.21
-        , testAFormat "%yy%Bm%ww%Dd%Hh%Mm%Ss" "-5y-4m-3w-2d-2h-22m-8s" $
-            CalendarDiffTime (-64) $
-                negate $
-                    23 * 86400 + 8528.21
-        , testAFormat "%yy%Bm%ww%Dd%Hh%Mm%ESs" "-5y-4m-3w-2d-2h-22m-8.21s" $
-            CalendarDiffTime (-64) $
-                negate $
-                    23 * 86400 + 8528.21
-        , testAFormat "%bm %dd %hh %mm %Ess" "-64m -23d -554h -33262m -1995728.21s" $
-            CalendarDiffTime (-64) $
-                negate $
-                    23 * 86400 + 8528.21
-        ]
--}
 
 testFormat :: TestTree
 testFormat =
@@ -221,6 +141,6 @@ testFormat =
       testTimeZone,
       testNominalDiffTime,
       testDiffTime,
-      testCalenderDiffDays,
-      testCalenderDiffTime
+      testCalendarDiffDays,
+      testCalendarDiffTime
     ]